Refactor capture: add interface, Windows impl
add core types (pkg/types/core.go). Several legacy files removed or relocated to reflect the refactor and othar stuff

+++ pkg/server/server.go
@@ -396,7 +396,7 @@ func (s *Server) handleAppDetails(w http.ResponseWriter, r *http.Request) {
 			FROM manual_notes
 			WHERE session_id = ?
 			ORDER BY created_at ASC
-		`, session.ID)
+		`, int64(session.ID))
 		if err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 			return
@@ -552,7 +552,7 @@ func (s *Server) handleSessionUpdate(w http.ResponseWriter, r *http.Request, dat
 		return
 	}
 
-	if err := s.upsertManualNotes(session.ID, metadata.ManualNotes); err != nil {
+	if err := s.upsertManualNotes(int64(session.ID), metadata.ManualNotes); err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
